+++ test/support/generic_helper.rb
@@ -29,7 +29,7 @@ class Minitest::HooksSpec
   end
 
   def image
-    File.open("test/fixtures/image.jpg")
+    File.open("test/fixtures/image.jpg", "rb")
   end
 
   def io?(object)